Added process worker pool
Added the `ProcessPool` which is a worker pool for processes. The use
case for this is when running eBPF in userspace against a packet capture
for example. A common technique in eBPF programs is to use per-CPU maps
these maps have seperate storage for each logical CPU so it is
physically impossible to access the same memory at the same time.
Running processes in a `ProcessPool` guarantees that processes with the
same CPUID will never be executed at the same time and that execution of
a process happens on the same "virtual CPU".

To communicate the virtual CPU on which a process runs, a new `CPUID`
field was added to the process. This field may be changed between each
CPU instruction to indicate is has been "rescheduled", doing so is up
the the way the Host code schedules the processes.