Added per-cpu array map support
Added the per-cpu array map. Updated the `LinuxMap` and
`LinuxMapUpdater` interfaces to get calles to include the CPU ID for
which to do the operation. Since a significat amount of maps are
per-CPU, this feels reasonable. Normal maps can just ignore the
parameter.

The PerCPU array is implemented much like the ProgArray variant where it
just includes a slice of normal arrays, this avoids code-duplication.
The only downside is that this is less memory efficient, a price I am
currently willing to pay.