Fixed bug in fixupJumpsAndCalls
There was a bug in the fixupJumpsAndCalls which was copied from the ebpf
project. In ebpf they use an iterator to translate from raw offsets to
actual instruction offsets, since we have to often perform offset to
instruction traslations, we add a fake "Nop" instruction after
instruction which actually take up 2 instructions worth of space. Using
both the "Nop"s and the iterator caused issues, so replaced the iterator
with just a normal for-each