Added array-like and hash-like maps to MapSpecToLinuxMap
Some maps are unqique in the way they function, like stack, queues and
ringbuffers. However, most of the "specialized" maps are just array maps
or hash maps with very specific restrictions on the type of the pointer
stored in its value.

For us, as emulator, there is no need to implement every type explicitly
as long as the behavour from the eBPF VM's perspective is consitant with
the Linux kernel. Since the memory controller allows us to reference any
type and resolve it in Go/the helpers, there is no need for actual
seperate map implementations.