Added RingMemory and split up the memory_controller file
There are a few map types which profit from using ring buffers as
underlaying storage. The RingMemory makes implementation of ring buffers
easier. It is very similar to PlainMemory except when partially reading
out of bounds(`offset < len(buf) && offset+size >= len(buf)`) the buffer
will wrap around and read the remaining bytes from the start of the
memory as if it was an actual ring. So any offset between 0 and len(buf)
is valid regardless of the size.