Added BPF_MAP_TYPE_PERF_EVENT_ARRAY support
Added a new map type LinuxPerfEventArrayMap which emulates the perf
event array map. Normally in linux a loader would create a perf event
and associate ring buffers from the userspace. The emulator uses a
setting on the LinuxEmulator instead.

The perf event array is a pseudo per-CPU map, the MaxEntries setting on
the map Spec can be used to set the amount of rings, if 0 the amount of
vCPUs is used. The program can specify which ring to used using the
flags passed to the bpf_perf_event_output helper or let use its current
CPU id by passing in BPF_F_CURRENT_CPU.

Perf events can be of variable size, our emulated implementation
therefor prepends a uint32 with the size of the data. This does mean
that some of the ringbuffer is used for this size tracking.