Added queue and stack map support
Added the LinuxQueueMap and LinuxStackMap types which are a FIFO queue
and FILO stack respectively. Also added the bpf_map_push_elem,
bpf_map_pop_elem, and bpf_map_peek_elem helper functions which are used
to interact with these two map types.

The ringBuffer was moved into its own file since the queue now also uses
it, not just the perf_event_array.

For some reason, linux implements stacks and queues in such a way that
the peek and pop operations will return pointers into the maps. So where
in the normal lookup we would write a value directly to the provided
value pointer, with these map types we write the address to the value
to the provided value pointer and the program then will access it
outside of the spinlock/mutex protection of the queue/stack. This seems
to be a performance optimization, assuming the program will read its
value before the references memory is re-used. Because of this change,
the perf_event_buffer doesn't fit with the LinuxMapPopper interface
anymore, and reading from it should be handled seperately.