Added socket buffer context
Added the SKBuff object which is a bad replica of the Linux kernel's
sk_buff structure. It only includes the fields which can be accessed via
the __sk_buff proxy struct. As we add more helpers in the future this
struct may grow.

Currently, the only way to construct a SKBuff is from a byte slice which
will be parsed as a network packet. In the future, the
LinuxContextSKBuff might include more fields which can be used to set
specific values in the SKBuff.