Added captured context support
Added support for captured contexts which will allow `edb` to replay
actual program runs which were captured in the kernel.

Some helpers can be replayed, some can't. The readme contains an
overview of which helpers are supported in which "modes".
The LinuxEmulator will check if the current context is a CapturedContext
if so, and the context contains a response for the helper call, we will
replay it. If replaying is not possible, emulation is attempted, which
for some calls is not possible in which case an error will be thrown.